Address Copilot PR review
- Break the contest.mapper / contest.participant.mapper circular
  dependency: mapContestParticipantDto now lives in
  contest.participant.mapper alongside mapParticipantDashboardDto;
  contest.mapper re-exports both. participant mapper no longer
  imports from contest.mapper.
- Type the list endpoints in examQuestions.repository.ts as
  ExamQuestionDto[] (was ExamQuestionDto / unknown), matching the
  Array.isArray runtime branch.
